I just added the whole folder so you guys can also access the files
just download the latest version and go from that! make sure you let other people know when you are currently making changes so not many people do changes on the same version! I added the red tiles js as a guide for where we are going to put the tiles (tile level)